# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notes Doc API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/": {},
        "/note/": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Список",
                "operationId": "1c766328215231aca4fb72c1fc4d7ecb",
                "parameters": [
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Фильтр по дате создания записки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2025-06-07T00:00:00Z"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество записок на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK, возвращает HTML-страницу с записями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "html": {
                                            "type": "string",
                                            "example": "<html>...</html>"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Создание",
                "operationId": "408303f15d070ebb82bc2df7a8795ef8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "note": {
                                                "type": "string",
                                                "example": "Text note"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "note": {
                                            "type": "string",
                                            "example": "Text note"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Notes",
            "description": "Notes"
        }
    ]
}